fix: added login on hdr & tested fr responsiveness
