+++ client/src/components/header/header.tsx
@@ -1,5 +1,6 @@
 import Nav from "./nav";
 import FFTLogo from "../../assets/images/logo.png";
+import LoginForm from "../login_form/login_form";
 
 const Header: React.FC = () => (
   <header className="bg-yellow-400 text-green-900 border border-black p-2 font-bold w-full flex items-center justify-center box-border">
@@ -9,6 +10,7 @@ const Header: React.FC = () => (
       className="block mx-auto h-32 m-2"
     />
     <Nav />
+    <LoginForm />
   </header>
 );
 export default Header;
